chore: move team activity methods to a separate module
