+++ app/models/concerns/activity_message_handler.rb
@@ -1,10 +1,10 @@
-# rubocop:disable Metrics/ModuleLength
 module ActivityMessageHandler
   extend ActiveSupport::Concern
 
   include PriorityActivityMessageHandler
   include LabelActivityMessageHandler
   include SlaActivityMessageHandler
+  include TeamActivityMessageHandler
 
   private
 
@@ -94,30 +94,6 @@ module ActivityMessageHandler
     ::Conversations::ActivityMessageJob.perform_later(self, activity_message_params(content)) if content
   end
 
-  def generate_team_change_activity_key
-    team = Team.find_by(id: team_id)
-    key = team.present? ? 'assigned' : 'removed'
-    key += '_with_assignee' if key == 'assigned' && saved_change_to_assignee_id? && assignee
-    key
-  end
-
-  def generate_team_name_for_activity
-    previous_team_id = previous_changes[:team_id][0]
-    Team.find_by(id: previous_team_id)&.name if previous_team_id.present?
-  end
-
-  def create_team_change_activity(user_name)
-    user_name = activity_message_ownner(user_name)
-    return unless user_name
-
-    key = generate_team_change_activity_key
-    params = { assignee_name: assignee&.name, team_name: team&.name, user_name: user_name }
-    params[:team_name] = generate_team_name_for_activity if key == 'removed'
-    content = I18n.t("conversations.activity.team.#{key}", **params)
-
-    ::Conversations::ActivityMessageJob.perform_later(self, activity_message_params(content)) if content
-  end
-
   def generate_assignee_change_activity_content(user_name)
     params = { assignee_name: assignee&.name, user_name: user_name }.compact
     key = assignee_id ? 'assigned' : 'removed'
@@ -138,15 +114,4 @@ module ActivityMessageHandler
     user_name = 'Automation System' if !user_name && Current.executed_by.present?
     user_name
   end
-
-  def determine_sla_change_type
-    sla_policy_id_before, sla_policy_id_after = previous_changes[:sla_policy_id]
-
-    if sla_policy_id_before.nil? && sla_policy_id_after.present?
-      'added'
-    elsif sla_policy_id_before.present? && sla_policy_id_after.nil?
-      'removed'
-    end
-  end
 end
-# rubocop:enable Metrics/ModuleLength
